fix(icon): never draw an icon whose bitmap failed to load
OpenMeteo code 82 was converted to iconId 530, which no icon pack contains and
WEATHER_CODE_MAPPING does not map. getIconPath then built a path to a missing
`null.png`, loadImageSync wrapped the null bitmap into a truthy ImageSource, and
canvas.drawBitmap crashed with a native NPE on every frame, so the app could not
draw its main page any more.

Convert code 82 to 522, map 530 for data already cached, return null instead of a
path to a missing file, and drop an ImageSource that has no underlying bitmap.
Extract the icon id resolution and the OpenMeteo code conversion into pure
modules so the mapping is covered by tests.

Fixes #504
